+++ src/components/GradDemands/index.jsx
@@ -4,33 +4,33 @@ import React, { Component } from "react";
 import { Accordion, Card } from "react-bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
 
-import icon1 from './Assets/numbericons/1.png';
-import icon2 from './Assets/numbericons/2.png';
-import icon3 from './Assets/numbericons/3.png';
-import icon4 from './Assets/numbericons/4.png';
-import icon5 from './Assets/numbericons/5.png';
-import icon6 from './Assets/numbericons/6.png';
-import icon7 from './Assets/numbericons/7.png';
-import icon8 from './Assets/numbericons/8.png';
-import icon9 from './Assets/numbericons/9.png';
-import icon10 from './Assets/numbericons/10.png';
-import icon11 from './Assets/numbericons/11.png';
-import icon12 from './Assets/numbericons/12.png';
-import icon13 from './Assets/numbericons/13.png';
-import icon14 from './Assets/numbericons/14.png';
-import icon15 from './Assets/numbericons/15.png';
-import icon16 from './Assets/numbericons/16.png';
-import icon17 from './Assets/numbericons/17.png';
-import icon18 from './Assets/numbericons/18.png';
-import icon19 from './Assets/numbericons/19.png';
-import icon20 from './Assets/numbericons/20.png';
-import icon21 from './Assets/numbericons/21.png';
-import icon22 from './Assets/numbericons/22.png';
-import icon23 from './Assets/numbericons/23.png';
-import icon24 from './Assets/numbericons/24.png';
-import icon25 from './Assets/numbericons/25.png';
+import icon1 from '../../Assets/numbericons/1.png';
+import icon2 from '../../Assets/numbericons/2.png';
+import icon3 from '../../Assets/numbericons/3.png';
+import icon4 from '../../Assets/numbericons/4.png';
+import icon5 from '../../Assets/numbericons/5.png';
+import icon6 from '../../Assets/numbericons/6.png';
+import icon7 from '../../Assets/numbericons/7.png';
+import icon8 from '../../Assets/numbericons/8.png';
+import icon9 from '../../Assets/numbericons/9.png';
+import icon10 from '../../Assets/numbericons/10.png';
+import icon11 from '../../Assets/numbericons/11.png';
+import icon12 from '../../Assets/numbericons/12.png';
+import icon13 from '../../Assets/numbericons/13.png';
+import icon14 from '../../Assets/numbericons/14.png';
+import icon15 from '../../Assets/numbericons/15.png';
+import icon16 from '../../Assets/numbericons/16.png';
+import icon17 from '../../Assets/numbericons/17.png';
+import icon18 from '../../Assets/numbericons/18.png';
+import icon19 from '../../Assets/numbericons/19.png';
+import icon20 from '../../Assets/numbericons/20.png';
+import icon21 from '../../Assets/numbericons/21.png';
+import icon22 from '../../Assets/numbericons/22.png';
+import icon23 from '../../Assets/numbericons/23.png';
+import icon24 from '../../Assets/numbericons/24.png';
+import icon25 from '../../Assets/numbericons/25.png';
 
-import plex from './fonts/plexSans.ttf';
+import plex from '../../fonts/plexSans.ttf';
 
 let accordionPadding = '10%';
 let iconWidth = '75%';
